pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ku-secretary"
version = "0.1.0"
description = "KU Secretary local-first sync agent"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "KU Secretary" }]
dependencies = [
  "typer>=0.12.3",
  "python-dotenv>=1.0.1",
  "requests>=2.31.0",
  "icalendar>=5.0.12",
  "python-dateutil>=2.9.0",
  "olefile>=0.47",
  "python-pptx>=1.0.2",
  "pypdf>=4.3.1",
  "playwright>=1.52.0",
  # KUPID/Canvas connectors (vendored from ku-portal-mcp v0.10.1)
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.0.0",
  "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[project.scripts]
kus = "ku_secretary.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
ku_secretary = ["ops_dashboard_assets/*.html"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
markers = [
  "beta_critical: hermetic tests for the Korea University beta-critical path",
]
